+++ tests/system/test_asyncpg_iam_auth.py
@@ -17,9 +17,9 @@ limitations under the License.
 import asyncio
 import os
 
-import asyncpg
 import sqlalchemy
 import sqlalchemy.ext.asyncio
+import logging
 
 from google.cloud.sql.connector import Connector
 
@@ -64,21 +64,19 @@ async def create_sqlalchemy_engine(
     loop = asyncio.get_running_loop()
     connector = Connector(loop=loop, refresh_strategy=refresh_strategy)
 
-    async def getconn() -> asyncpg.Connection:
-        conn: asyncpg.Connection = await connector.connect_async(
+    # create SQLAlchemy connection pool
+    engine = sqlalchemy.ext.asyncio.create_async_engine(
+        "postgresql+asyncpg://",
+        async_creator=lambda: connector.connect_async(
             instance_connection_name,
             "asyncpg",
             user=user,
             db=db,
-            ip_type="public",  # can also be "private" or "psc"
+            ip_type=os.environ.get(
+                "IP_TYPE", "public"
+            ),  # can be "public","private" or "psc"
             enable_iam_auth=True,
-        )
-        return conn
-
-    # create SQLAlchemy connection pool
-    engine = sqlalchemy.ext.asyncio.create_async_engine(
-        "postgresql+asyncpg://",
-        async_creator=getconn,
+        ),
         execution_options={"isolation_level": "AUTOCOMMIT"},
     )
     return engine, connector
@@ -112,3 +110,7 @@ async def test_lazy_iam_authn_connection_with_asyncpg() -> None:
         assert res[0] == 1
 
     await connector.close_async()
+
+logging.basicConfig(format="%(asctime)s [%(levelname)s]: %(message)s")
+logger = logging.getLogger(name="google.cloud.sql.connector")
+logger.setLevel(logging.DEBUG)