+++ tests/system/test_pytds_connection.py
@@ -1,4 +1,4 @@
-""""
+"""
 Copyright 2021 Google LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +17,6 @@ limitations under the License.
 import os
 
 # [START cloud_sql_connector_mysql_pytds]
-import pytds
 import sqlalchemy
 
 from google.cloud.sql.connector import Connector
@@ -65,21 +64,19 @@ def create_sqlalchemy_engine(
     """
     connector = Connector(refresh_strategy=refresh_strategy)
 
-    def getconn() -> pytds.Connection:
-        conn: pytds.Connection = connector.connect(
+    # create SQLAlchemy connection pool
+    engine = sqlalchemy.create_engine(
+        "mssql+pytds://",
+        creator=lambda: connector.connect(
             instance_connection_name,
             "pytds",
             user=user,
             password=password,
             db=db,
-            ip_type="public",  # can also be "private" or "psc"
-        )
-        return conn
-
-    # create SQLAlchemy connection pool
-    engine = sqlalchemy.create_engine(
-        "mssql+pytds://",
-        creator=getconn,
+            ip_type=os.environ.get(
+                "IP_TYPE", "public"
+            ),  # can be "public","private" or "psc"
+        ),
     )
     return engine, connector